Dump full host memory when no PIDs supplied
Up until now the application has been designed to run on a host and
pointed at a process (guest VM) to analyze. For in-guest page tacking we
want to rather look at the whole local address space. This will be done
if no PIDs are supplied as cmd line args, and the resuling file will
have the virtual address set to 0x0.