Only set/read idlemap bytes for RAM pages
There are more PFNs than just system RAM. Up until now the idlemap
getting and setting and set and read every page. This change first looks
for which address ranges are System RAM, and only sets bits for the PFNs
in System RAM ranges. The code is a bit of a mess with all of the
rounding up and rounding down and setting to 8-byte boundaries.