Introduce QEMU pageout for idle pages
A sample of pages which are idle are passed to QEMU via QMP JSON message
and QEMU will page these pages out.

To do this, a Vmm struct is introduced which represents an interface to
the Virtual Machine Monitor (QEMU). It is responsibile for establishing
the QPM telnet session and marahaling pages addresses via JSON to QEMU.
Currently this Rust app doesn't integrate with QEMU header files, but it
probably could be made to in future.