Write trackee memory stats to CSV
First stab at getting some data written out on each observation loop to
a CSV about what the process is up to.
Objective is to feed this CSV into a graph plotter so we can observe
trends over time without needing to parse the whole kpageflats array.
In fact, unless we actually want to look at memory activity on a
per-page basis we can avoid persisting and uploading the full kpageflags
array now.