Log swapped out data size
We want this so we can know what percentage of the full guest size we've
managed to swap out.
Yucky hack reading /proc/pid/smaps and "grepping" for the Swap line.
Really this should come from the sysinfo crate, and I'll look at
extending it in the near future to know about segments.